requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-rest-auth==2.1.3
Django==3.1.7
django-allauth==0.44.0
django-braces==1.14.0
django-constance==2.8.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-polymorphic==3.0.0
django-rest-auth==0.9.5
django-rest-registration==0.5.6
django-sslserver==0.22
djangorestframework==3.12.2
djangorestframework-datatables==0.5.2
drf-base64==2.0
drf-yasg==1.20.0
face-recognition-models==0.3.0
html5lib==1.0.1
httplib2==0.19.0
idna==2.10
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.19.4
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
pep517==0.8.2
Pillow==8.0.1
progress==1.5
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.1.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
webencodings==0.5.1
